+++ common/src/main/java/net/mehvahdjukaar/polytone/mixins/SubmitNodeStorageMixin.java
@@ -4,6 +4,7 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import net.mehvahdjukaar.polytone.Polytone;
 import net.minecraft.client.model.Model;
 import net.minecraft.client.renderer.SubmitNodeStorage;
+import net.minecraft.client.renderer.entity.state.EntityRenderState;
 import net.minecraft.client.renderer.feature.ModelFeatureRenderer;
 import net.minecraft.client.renderer.rendertype.RenderType;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
@@ -23,7 +24,8 @@ public class SubmitNodeStorageMixin {
                                                                       int l, ModelFeatureRenderer.@Nullable CrumblingOverlay crumblingOverlay,
                                                                       CallbackInfo ci) {
         //Cant use forge events as they are missing the camera state
-
-        Polytone.ENTITY_MODIFIERS.onEntityRender(model , poseStack, object);
+        if (object instanceof EntityRenderState state) {
+            Polytone.ENTITY_MODIFIERS.onEntityRender((Model<? super EntityRenderState>) model, poseStack, state);
+        }
     }
 }